requirements.txt
# Core Dependencies
discord.py
python-dotenv
google-genai
requests

# DSPy Framework
dspy-ai

# Pinecone Vector DB (Updated name)
pinecone

# AI/ML Services
groq
google-auth
google-api-python-client
google-generativeai

# Utilities
patool
gitpython
